requirements.txt
jupyter
torch
transformers
pillow
matplotlib
aiohttp
datasets